refactor: Extract ranking generator
